Fix QR Code Duplicate Import Warnings
- Merged qr-canvas.tsx and qr-code.tsx into single qr-code.tsx component
- Removed duplicate qr-generator.ts utility (kept qr-code.ts with ISO/IEC 18004 implementation)
- Both QRCanvas and QRCode components now exported from single file
- Resolves build warnings about duplicate imports
